feat: add filtering and sorting to list page
